+++ lib/active_record_proxy_adapters/connection_handling.rb
@@ -3,3 +3,4 @@
 require "active_record_proxy_adapters/connection_handling/postgresql"
 require "active_record_proxy_adapters/connection_handling/mysql2"
 require "active_record_proxy_adapters/connection_handling/trilogy"
+require "active_record_proxy_adapters/connection_handling/sqlite3"