+++ spec/active_record_proxy_adapters/sqlite3_proxy_spec.rb
@@ -1,17 +1,14 @@
 # frozen_string_literal: true
 
 require "shared_examples/a_proxied_method"
-require "shared_examples/a_transaction_block_proxy_bypass"
 
-RSpec.shared_examples "a MySQL proxy" do
+RSpec.describe ActiveRecordProxyAdapters::SQLite3Proxy do # rubocop:disable RSpec/SpecFilePathFormat
   attr_reader :primary_adapter
 
-  let(:replica_pool) { nil }
-  let(:primary_pool) { nil }
-  let(:adapter_class) { nil }
-  let(:model_class) { nil }
-
-  let(:truncate_database) { nil }
+  let(:replica_pool) { TestHelper.sqlite3_replica_pool }
+  let(:primary_pool) { TestHelper.sqlite3_primary_pool }
+  let(:adapter_class) { ActiveRecord::ConnectionAdapters::SQLite3Adapter }
+  let(:model_class) { TestHelper::SQLite3Record }
 
   around do |example|
     primary_pool.with_connection do |connection|
@@ -22,7 +19,7 @@ RSpec.shared_examples "a MySQL proxy" do
       @primary_adapter = nil
     end
 
-    truncate_database
+    TestHelper.truncate_sqlite3_database
   end
 
   def create_dummy_user
@@ -32,8 +29,6 @@ RSpec.shared_examples "a MySQL proxy" do
     SQL
   end
 
-  it_behaves_like "a transaction block proxy bypass"
-
   describe "#execute" do
     it_behaves_like "a proxied method", :execute
   end
@@ -42,7 +37,7 @@ RSpec.shared_examples "a MySQL proxy" do
     it_behaves_like "a proxied method", :exec_query
   end
 
-  if TestHelper.active_record_context.active_record_v7_1_or_greater?
+  if TestHelper.active_record_context.active_record_v8_0_or_greater?
     describe "#internal_exec_query" do
       it_behaves_like "a proxied method", :internal_exec_query
     end